04 Neo4J/Python/backend/requirements.txt
Flask==2.0.1
neo4j==4.3.1
PyJWT==2.1.0
Werkzeug==2.0.1
Flask-Cors==3.0.10